meta-imx: reduce layer priority to 5
Issue: LINUXEXEC-34713

Reduce the priority of meta-imx, because if the priority
of meta-imx is higher, it may cause the package in
meta-imx to be preferred during compilation. Since the
update frequency of meta-imx is slow, so the version of the
package may be stale, which may lead to compilation fail.

(LOCAL REV: NOT UPSTREAM) - WRLinux specific

Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>