meta-imx: add condition check for openssl-no-weak-ciphers
Issue: LIN1025-2982

When the template openssl-no-weak-ciphers is enabled, "no-des" is added
to openssl's EXTRA_OECONF. This will result that the package
wpa-supplicant can't be built with "openssl" support. So add condition
check to ensure that wpa-supplicant can only be built with "openssl"
support under no openssl-no-weak-ciphers condition.

(LOCAL REV: NOT UPSTREAM) - WRLinux specific

Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>